[ES] Improve area expansion to support for masculine areas (#436)
I added some test to assert that turning on and off lights of areas with a masculine name (e.g. comedor) worked, to make sure it picked both "del <area>" and `de el <area>`.

While the second is mildly incorrect (it should be contracted to "del"), it's a very common mispronunciation, specially if the user is doubts for a split second while speaking about the area's name.