[de-CH] Update files to consider additional dialects (#896)
* I updated the de-CH files to consider also some
versions as used in the bernese dialect (e.g."aui").
Furthermore I removed variances I think that were
considered double (once in HassTurnOn and once in
the expansion rules).
Furthermore, I started with some turn off sentences.
Finally, I added some test sentences.
The commit did not yet pass all tests (not all sentences have
tests), as in the test-file quite some additional variances have
to be added.

* Amended files based on comments received
Tests are still not passed

* added changes in _common.yaml

* added response

* responses belongs to `sentences` not to `slots:`

* Amended de-CH files to pass tests

I finally got all tests passed. Had to amend some sentences.
In this course I also added a differentiation to turn on/off one
light or all lights in an area (with different response).
This should be a good basis for further fine tuning.

* [de-CH] Add sentences for HassLightSet and fan
I added translations for the sentences for HassLightSet and
fan_HassTurnOn and fan_HassTurnOff (including the responses and tests).

* undo unwanted restriction added for debugging

* [de-ch] rename languages
According to my understanding, the native name in languages.yaml
should be in the local language. I propose one spelling, feel free
to change to other dialect (but Swiss-German definitively is not
native language).

* added "" in response-files

* Fix fans sentences

* add "requires_context" for light

---------

Co-authored-by: schizza <lukyn@lukyns.com>