Updating Hungarian intentions
The current intentions are unusable, as they do not employ conjugations at any level. The entire language, which utilizes conjugations, has been updated.

Due to the nature of the Hungarian language, there's an issue where word roots change with conjugation. (For example, if the room is a kitchen, instead of "kapcsold be az összes ventilátort a konyhaban", it should be "konyhában".)

I circumvented these with suffix words, like using "piros színre" instead of "pirosra", or by changing the word order to avoid these issues. This way, it works, like "kapcsold be a konyha összes ventilátorát is."

But for correct and complete Hungarian expressions, it's advisable to use assistant aliases for room names whose endings change with conjugation, such as "szoba" (room), "konyha" (kitchen). For these, "szobá" and "konyhá" aliases should be added.