[HU] Optimization of response, handling of articles, correction of number format
I have corrected the responses to put the correct article in front of the word.
The number format correction involves changing 12.1 to 12,1 and replacing "-" with the word "minus".

And an attempt to find out what cover states the live system returns in the live environment. If it's English, then from now on it should also display it in Hungarian; if it's not English, then everything remains as it was.