[HU] Extracting optional variables from the variations (#1921)
* [HU] Migrate errors

* [hu] error fix

It seems that the temperature adjustment does not work in the beta. It conflicts with the scene command. I made the use of the words "scene" and "script" mandatory for the scene and the script. I hope it solves the issue.

* [HU] color management correction

I simplified the color setting due to fixing false matches.
          - "([<set>];<name> [<light>];szín(űre|e|re|ét);{color})"

The word for 'szín' 'color is required from now on.

* [HU] Extracting optional variables from the variations