Add new sentence to set a light to a specific brightness
I just started controlling my home with my voice using an Onju Home a couple days ago and I noticed one sentence that I used with alexa ALL THE TIME that wasn't recognized.

The ability to say "Enciende la luz de la cocina al 30%", as 'Enciende' was not recognized as a way of triggering setting a light to a specific brightness, but I think it's a very natural phase, specially when the lights are off.