[en] Tighten the matching for general turn on and off statements
This will prevent ambiguous sentences from being recognized as an turn on or off statement.

Eg. If you currently say 'Is the unknown device on', the '<area> <name> on' sentence was matched and
the user recieved a confusing error suggesting there is no area called 'Is'. This will require an
explicit request to 'turn|switch|etc' something on or off.