[DE] more options to cancel a timer (#2984)
* Give more options to cancel a Timer in German

The Imperative of stopp and lösch has two viable options, with and without an e.
 
The new word option lösch(e) isn't already used in any other common option, and is German for delete. So there shouldn't be any conflicts created by this change. 

It is also a variant to cancel a timer that is allowed on Google Home and possibly Alexa (I don't have one).
Allowing it makes switching away from them easier, because you do not have to learn a now command word for the same action.

* [DE] added tests for commit 2984