[FR] weather state is now exposed raw and untranslated (#2910)
* [FR] weather state is now exposed raw and untranslated

As it goes for every exposed entity of any domain from now on.

This concludes uncertainty about the content of "state.state", and will
have the nice side effect of putting the development environment on par
with the production behaviour.

Fixes #2899

* [FR] ignore when wind gust speed is exactly zero

When protecting against a None value, I let that one slip through. Test
for any false value.